requirements.txt
Flask==2.1.0
google-cloud-bigquery==3.17.2
livereload==2.6.3
matplotlib==3.5.2
numpy==1.24.1
nba_api==1.4.1
pandas==1.4.4
pandas-gbq==0.21.0
psycopg2==2.9.9
python-dotenv==0.20.0
scikit-learn==1.4.1.post1
sqlalchemy==2.0.27
typing==3.7.4.3
Werkzeug==3.0.3